refactors and cleanup - create "data" package - delete "networking" package - clean up initialisation of tracked data   - delete ColourData.java   - migrate functionality to Rainglow.java - delete unused "unlock config" packet
