+++ src/main/java/io/ix0rai/rainglow/config/RainglowConfigScreen.java
@@ -6,8 +6,8 @@ import dev.lambdaurora.spruceui.option.SpruceOption;
 import dev.lambdaurora.spruceui.option.SpruceSimpleActionOption;
 import dev.lambdaurora.spruceui.widget.SpruceLabelWidget;
 import io.ix0rai.rainglow.Rainglow;
-import io.ix0rai.rainglow.RainglowMode;
-import io.ix0rai.rainglow.SquidColour;
+import io.ix0rai.rainglow.data.RainglowMode;
+import io.ix0rai.rainglow.data.SquidColour;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.client.toast.SystemToast;
