+++ src/main/java/io/ix0rai/rainglow/mixin/client/GlowSquidEntityRendererMixin.java
@@ -1,8 +1,7 @@
 package io.ix0rai.rainglow.mixin.client;
 
 import io.ix0rai.rainglow.Rainglow;
-import io.ix0rai.rainglow.SquidColour;
-import io.ix0rai.rainglow.ColourData;
+import io.ix0rai.rainglow.data.SquidColour;
 import net.minecraft.client.render.entity.GlowSquidEntityRenderer;
 import net.minecraft.entity.passive.GlowSquidEntity;
 import net.minecraft.util.Identifier;
@@ -19,7 +18,7 @@ public class GlowSquidEntityRendererMixin {
      */
     @Inject(method = "getTexture*", at = @At("HEAD"), cancellable = true)
     public void getTexture(GlowSquidEntity glowSquidEntity, CallbackInfoReturnable<Identifier> cir) {
-        String colour = ColourData.getColour(glowSquidEntity.getDataTracker(), glowSquidEntity.getRandom());
+        String colour = Rainglow.getColour(glowSquidEntity.getDataTracker(), glowSquidEntity.getRandom());
 
         // if the colour is blue we don't need to override the method
         // this optimises a tiny bit
